Fix expected/actual order in all tests
Well, that was tedious. Thank goodness for vim-mode macros. I realized while writing the tests for the hot potato queue example that assertEquals(actual, expected) considers first parameter to be the thing under test (actual), and the second to be the expected result. I had it reversed, for reasons unknown. Worth fixing, since the output of a failing test uses that definition.