Test and properly type stack and queue constructors
The book does not include convenience constructors (for making a stack, queue, or dequeue from an array), but in the palindrome checker, it uses this functionality from the python collections.deque implementation. I had already implemented this, but neglected to test it. Testing revealed it was improperly typed. It's handy, so I'll keep it in.